lint: clear the golangci-lint v2 baseline (31 → 0)
Resolves every issue `make lint` surfaced against the v2 preset so the
CI lint job can flip from advisory to blocking in a follow-up.

- `.golangci.yml`: disable `gocritic/rangeValCopy` globally (noisy on
  `map[string]config.ToolDefinition` iteration; none of the sites are
  hot paths) and extend the `internal/bridge/` path exclusion to
  `govet` so the cgo callback-id idiom
  (`unsafe.Pointer(uintptr(id))`, ~20 call sites) stops tripping
  `unsafeptr`.
- `nilerr`: add targeted `//nolint:nilerr` on 5 `filepath.Walk`
  callback returns where skipping walk errors is intentional; rename
  one local shadow (`ierr` in engine/pull_bytes.go).
- `revive/unused-parameter`: rename unused `cmd *cobra.Command` /
  `args []string` params to `_` across ~20 cobra handlers and 3 test
  mocks.
- `gocritic/importShadow`: rename local `hex` shadows in
  cache/rootfs.go and engine/runtime_test.go.
- `gocritic/appendCombine`: collapse a 3-append chain in
  commands/clean.go.
- `staticcheck/SA4010`: remove dead `choices` slice in
  commands/init.go.
- `unused`: drop the unused `cleanBucket.doRemove` field.
- `unconvert`: drop `uint64()` / `int64()` casts on darwin types that
  are already the target width (Termios.*flag, Stat_t.Blocks).

+++ cmd/silo/main.go
@@ -165,7 +165,7 @@ func resetTerminalIfNeeded() {
 	wantL := uint64(unix.ICANON | unix.ECHO | unix.ECHOE | unix.ISIG)
 	wantI := uint64(unix.ICRNL)
 	wantO := uint64(unix.OPOST)
-	if uint64(tios.Lflag)&wantL == wantL && uint64(tios.Iflag)&wantI == wantI && uint64(tios.Oflag)&wantO == wantO {
+	if tios.Lflag&wantL == wantL && tios.Iflag&wantI == wantI && tios.Oflag&wantO == wantO {
 		return
 	}
 	tios.Lflag |= unix.ICANON | unix.ECHO | unix.ECHOE | unix.ISIG
